Saving and Loading system finished
I finished the saving and loading system so that multiple topologies can be saved into one slot, from which they can be loaded from only using the name of the topology.

In addition, they can also be multiple slots so that you can have one slot for multiple topologies for one purpose and then have another for another purpose.

I have added regions to make it easier to read, and conformed to standard code formatting to make it easier to maintain