LeapBridge fully tested and BaseTopology
Tested and debugged my leap bridge functionality so can send leap data from the Leap Component to the data collector to the neural network with a single button click, or blueprint event call. One caveat being it is on the actor that all relevant components are attached to.

Also made a quick way to make a topology. All is needed is to add elements to the baseTopology array, with each element being a separate layer with the int value being the number of neurons in each layer. As it's used for initialisation, it'll have a default weight of 0 or a random value based on the relevant Boolean and it will be a fully connected network. This means that all neurons in one layer will be connected to all the neurons in the next layer

Plus, added in constructor for all neural network relevant USTRUCTS