Back Propagation #2 + Major Updates
Output weights and theta values are now training correctly. Still need to fine tune the code for hidden and input weights and theta values, but it's still progress.

Also added in some more safe guards against crashing. It will now print a message to the log and cancel out of whatever operation was occuring instead of crashing.

Threshold value is now on a per neuron basis.

Laid some groundwork for a network that isn't fully connected, so I can prune some connections later for optimisation and it'll still work. (Not tested as back propagation is my priority)

Encapsulated the activation and derivation functions so that they aren't reliant on a parameter to give them a threshold value.

Made an initialise function to replace what I was calling on BeginPlay, so that it is a separate and therefore more controllable operation.