Back Propagation #3 - Delta Weights
Finally made it to a milestone with my back propagation. The weights for each layer now adjust appropriately for each training session and in the direction that I expect them to.

Delta Theta is still an issue I have to deal with as it isn't being applied appropriately but that should be an easy fix for my next commit.